Move frequently reused profile+dataset setup/cleanup to base class.
