+++ datarepo-clienttests/src/main/java/testscripts/IngestFile.java
@@ -1,27 +1,23 @@
 package testscripts;
 
 import bio.terra.datarepo.api.RepositoryApi;
-import bio.terra.datarepo.api.ResourcesApi;
 import bio.terra.datarepo.client.ApiClient;
-import bio.terra.datarepo.model.BillingProfileModel;
 import bio.terra.datarepo.model.BulkLoadArrayRequestModel;
 import bio.terra.datarepo.model.BulkLoadArrayResultModel;
 import bio.terra.datarepo.model.BulkLoadFileModel;
 import bio.terra.datarepo.model.BulkLoadFileResultModel;
-import bio.terra.datarepo.model.DatasetSummaryModel;
-import bio.terra.datarepo.model.DeleteResponseModel;
 import bio.terra.datarepo.model.JobModel;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import testscripts.baseclasses.SimpleDataset;
 import utils.DataRepoUtils;
 import utils.FileUtils;
 
-public class IngestFile extends runner.TestScript {
+public class IngestFile extends SimpleDataset {
   private static final Logger logger = LoggerFactory.getLogger(IngestFile.class);
 
   /** Public constructor so that this class can be instantiated via reflection. */
@@ -30,9 +26,6 @@ public class IngestFile extends runner.TestScript {
   }
 
   private URI sourceFileURI;
-  private String datasetCreator;
-  private BillingProfileModel billingProfileModel;
-  private DatasetSummaryModel datasetSummaryModel;
 
   public void setParameters(List<String> parameters) throws Exception {
     if (parameters == null || parameters.size() == 0) {
@@ -45,33 +38,6 @@ public class IngestFile extends runner.TestScript {
         throw new RuntimeException("Error parsing source file URI: " + parameters.get(0), synEx);
       }
     logger.debug("Source file URI: {}", sourceFileURI);
-  }
-
-  public void setup(Map<String, ApiClient> apiClients) throws Exception {
-    // pick the first user to be the dataset creator
-    List<String> apiClientList = new ArrayList<>(apiClients.keySet());
-    datasetCreator = apiClientList.get(0);
-
-    // get the ApiClient for the dataset creator
-    ApiClient datasetCreatorClient = apiClients.get(datasetCreator);
-    ResourcesApi resourcesApi = new ResourcesApi(datasetCreatorClient);
-    RepositoryApi repositoryApi = new RepositoryApi(datasetCreatorClient);
-
-    // create a new profile
-    billingProfileModel =
-        DataRepoUtils.createProfile(resourcesApi, billingAccount, "profile-simple", true);
-    logger.info("Successfully created profile: {}", billingProfileModel.getProfileName());
-
-    // make the create dataset request and wait for the job to finish
-    JobModel createDatasetJobResponse =
-        DataRepoUtils.createDataset(
-            repositoryApi, billingProfileModel.getId(), "dataset-simple.json", true);
-
-    // save a reference to the dataset summary model so we can delete it in cleanup()
-    datasetSummaryModel =
-        DataRepoUtils.expectJobSuccess(
-            repositoryApi, createDatasetJobResponse, DatasetSummaryModel.class);
-    logger.info("Successfully created dataset: {}", datasetSummaryModel.getName());
   }
 
   public void userJourney(ApiClient apiClient) throws Exception {
@@ -105,24 +71,5 @@ public class IngestFile extends runner.TestScript {
         "Successfully ingested file: path = {}, id = {}",
         fileInfo.getSourcePath(),
         fileInfo.getFileId());
-  }
-
-  public void cleanup(Map<String, ApiClient> apiClients) throws Exception {
-    // get the ApiClient for the dataset creator
-    ApiClient datasetCreatorClient = apiClients.get(datasetCreator);
-    ResourcesApi resourcesApi = new ResourcesApi(datasetCreatorClient);
-    RepositoryApi repositoryApi = new RepositoryApi(datasetCreatorClient);
-
-    // make the delete request and wait for the job to finish
-    JobModel deleteDatasetJobResponse = repositoryApi.deleteDataset(datasetSummaryModel.getId());
-    deleteDatasetJobResponse =
-        DataRepoUtils.waitForJobToFinish(repositoryApi, deleteDatasetJobResponse);
-    DataRepoUtils.expectJobSuccess(
-        repositoryApi, deleteDatasetJobResponse, DeleteResponseModel.class);
-    logger.info("Successfully deleted dataset: {}", datasetSummaryModel.getName());
-
-    // delete the profile
-    resourcesApi.deleteProfile(billingProfileModel.getId());
-    logger.info("Successfully deleted profile: {}", billingProfileModel.getProfileName());
   }
 }
