+++ datarepo-clienttests/src/main/java/testscripts/RetrieveDataset.java
@@ -1,57 +1,18 @@
 package testscripts;
 
 import bio.terra.datarepo.api.RepositoryApi;
-import bio.terra.datarepo.api.ResourcesApi;
 import bio.terra.datarepo.client.ApiClient;
-import bio.terra.datarepo.model.BillingProfileModel;
 import bio.terra.datarepo.model.DatasetModel;
-import bio.terra.datarepo.model.DatasetSummaryModel;
-import bio.terra.datarepo.model.DeleteResponseModel;
-import bio.terra.datarepo.model.JobModel;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import utils.DataRepoUtils;
+import testscripts.baseclasses.SimpleDataset;
 
-public class RetrieveDataset extends runner.TestScript {
+public class RetrieveDataset extends SimpleDataset {
   private static final Logger logger = LoggerFactory.getLogger(RetrieveDataset.class);
 
   /** Public constructor so that this class can be instantiated via reflection. */
   public RetrieveDataset() {
     super();
-  }
-
-  private String datasetCreator;
-  private BillingProfileModel billingProfileModel;
-  private DatasetSummaryModel datasetSummaryModel;
-
-  public void setup(Map<String, ApiClient> apiClients) throws Exception {
-    // pick the first user to be the dataset creator
-    List<String> apiClientList = new ArrayList<>(apiClients.keySet());
-    datasetCreator = apiClientList.get(0);
-
-    // get the ApiClient for the dataset creator
-    ApiClient datasetCreatorClient = apiClients.get(datasetCreator);
-    ResourcesApi resourcesApi = new ResourcesApi(datasetCreatorClient);
-    RepositoryApi repositoryApi = new RepositoryApi(datasetCreatorClient);
-
-    // create a new profile
-    billingProfileModel =
-        DataRepoUtils.createProfile(resourcesApi, billingAccount, "profile-simple", true);
-    logger.info("Successfully created profile: {}", billingProfileModel.getProfileName());
-
-    // make the create dataset request and wait for the job to finish
-    JobModel createDatasetJobResponse =
-        DataRepoUtils.createDataset(
-            repositoryApi, billingProfileModel.getId(), "dataset-simple.json", true);
-
-    // save a reference to the dataset summary model so we can delete it in cleanup()
-    datasetSummaryModel =
-        DataRepoUtils.expectJobSuccess(
-            repositoryApi, createDatasetJobResponse, DatasetSummaryModel.class);
-    logger.info("Successfully created dataset: {}", datasetSummaryModel.getName());
   }
 
   public void userJourney(ApiClient apiClient) throws Exception {
@@ -61,24 +22,5 @@ public class RetrieveDataset extends runner.TestScript {
         "Successfully retrieved dataset: name = {}, data project = {}",
         datasetModel.getName(),
         datasetModel.getDataProject());
-  }
-
-  public void cleanup(Map<String, ApiClient> apiClients) throws Exception {
-    // get the ApiClient for the dataset creator
-    ApiClient datasetCreatorClient = apiClients.get(datasetCreator);
-    ResourcesApi resourcesApi = new ResourcesApi(datasetCreatorClient);
-    RepositoryApi repositoryApi = new RepositoryApi(datasetCreatorClient);
-
-    // make the delete dataset request and wait for the job to finish
-    JobModel deleteDatasetJobResponse = repositoryApi.deleteDataset(datasetSummaryModel.getId());
-    deleteDatasetJobResponse =
-        DataRepoUtils.waitForJobToFinish(repositoryApi, deleteDatasetJobResponse);
-    DataRepoUtils.expectJobSuccess(
-        repositoryApi, deleteDatasetJobResponse, DeleteResponseModel.class);
-    logger.info("Successfully deleted dataset: {}", datasetSummaryModel.getName());
-
-    // delete the profile
-    resourcesApi.deleteProfile(billingProfileModel.getId());
-    logger.info("Successfully deleted profile: {}", billingProfileModel.getProfileName());
   }
 }
