+++ datarepo-clienttests/src/main/java/testscripts/RetrieveSnapshot.java
@@ -1,32 +1,26 @@
 package testscripts;
 
 import bio.terra.datarepo.api.RepositoryApi;
-import bio.terra.datarepo.api.ResourcesApi;
 import bio.terra.datarepo.client.ApiClient;
-import bio.terra.datarepo.model.BillingProfileModel;
 import bio.terra.datarepo.model.BulkLoadArrayRequestModel;
 import bio.terra.datarepo.model.BulkLoadArrayResultModel;
 import bio.terra.datarepo.model.BulkLoadFileModel;
-import bio.terra.datarepo.model.DatasetSummaryModel;
 import bio.terra.datarepo.model.DeleteResponseModel;
 import bio.terra.datarepo.model.IngestRequestModel;
 import bio.terra.datarepo.model.IngestResponseModel;
 import bio.terra.datarepo.model.JobModel;
 import bio.terra.datarepo.model.SnapshotModel;
 import bio.terra.datarepo.model.SnapshotSummaryModel;
-import com.google.cloud.storage.Storage;
-import com.google.cloud.storage.StorageOptions;
 import java.net.URI;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import testscripts.baseclasses.SimpleDataset;
 import utils.DataRepoUtils;
 import utils.FileUtils;
 
-public class RetrieveSnapshot extends runner.TestScript {
+public class RetrieveSnapshot extends SimpleDataset {
   private static final Logger logger = LoggerFactory.getLogger(RetrieveSnapshot.class);
 
   /** Public constructor so that this class can be instantiated via reflection. */
@@ -34,42 +28,19 @@ public class RetrieveSnapshot extends runner.TestScript {
     super();
   }
 
-  private String datasetCreator;
-  private BillingProfileModel billingProfileModel;
-  private DatasetSummaryModel datasetSummaryModel;
   private SnapshotSummaryModel snapshotSummaryModel;
 
-  private Storage storage;
   private String testConfigGetIngestbucket;
 
   public void setup(Map<String, ApiClient> apiClients) throws Exception {
-    // pick the first user to be the dataset creator
-    List<String> apiClientList = new ArrayList<>(apiClients.keySet());
-    datasetCreator = apiClientList.get(0);
+    // create the profile and dataset
+    super.setup(apiClients);
 
-    // get the ApiClient for the dataset creator --- and the snapshot creator?
+    // get the ApiClient for the snapshot creator, same as the dataset creator
     ApiClient datasetCreatorClient = apiClients.get(datasetCreator);
-    ResourcesApi resourcesApi = new ResourcesApi(datasetCreatorClient);
     RepositoryApi repositoryApi = new RepositoryApi(datasetCreatorClient);
 
-    storage = StorageOptions.getDefaultInstance().getService();
     testConfigGetIngestbucket = "jade-testdata"; // this could be put in DRUtils
-
-    // create a new profile
-    billingProfileModel =
-        DataRepoUtils.createProfile(resourcesApi, billingAccount, "profile-simple", true);
-    logger.info("Successfully created profile: {}", billingProfileModel.getProfileName());
-
-    // make the create dataset request and wait for the job to finish
-    JobModel createDatasetJobResponse =
-        DataRepoUtils.createDataset(
-            repositoryApi, billingProfileModel.getId(), "dataset-simple.json", true);
-
-    // save a reference to the dataset summary model so we can delete it in cleanup()
-    datasetSummaryModel =
-        DataRepoUtils.expectJobSuccess(
-            repositoryApi, createDatasetJobResponse, DatasetSummaryModel.class);
-    logger.info("Successfully created dataset: {}", datasetSummaryModel.getName());
 
     // load data into the new dataset
     // note that there's a fileref in the dataset
@@ -152,7 +123,6 @@ public class RetrieveSnapshot extends runner.TestScript {
   public void cleanup(Map<String, ApiClient> apiClients) throws Exception {
     // get the ApiClient for the dataset creator
     ApiClient datasetCreatorClient = apiClients.get(datasetCreator);
-    ResourcesApi resourcesApi = new ResourcesApi(datasetCreatorClient);
     RepositoryApi repositoryApi = new RepositoryApi(datasetCreatorClient);
 
     // make the delete request and wait for the job to finish
@@ -163,19 +133,10 @@ public class RetrieveSnapshot extends runner.TestScript {
         repositoryApi, deleteSnapshotJobResponse, DeleteResponseModel.class);
     logger.info("Successfully deleted snapshot: {}", snapshotSummaryModel.getName());
 
-    // make the delete request and wait for the job to finish
-    JobModel deleteDatasetJobResponse = repositoryApi.deleteDataset(datasetSummaryModel.getId());
-    deleteDatasetJobResponse =
-        DataRepoUtils.waitForJobToFinish(repositoryApi, deleteDatasetJobResponse);
-    DataRepoUtils.expectJobSuccess(
-        repositoryApi, deleteDatasetJobResponse, DeleteResponseModel.class);
-    logger.info("Successfully deleted dataset: {}", datasetSummaryModel.getName());
+    // delete the profile and dataset
+    super.cleanup(apiClients);
 
     // delete scratch files
     FileUtils.cleanupScratchFiles(testConfigGetIngestbucket);
-
-    // delete the profile
-    resourcesApi.deleteProfile(billingProfileModel.getId());
-    logger.info("Successfully deleted profile: {}", billingProfileModel.getProfileName());
   }
 }
