+++ datarepo-clienttests/src/main/java/testscripts/ScalePodsToZero.java
@@ -5,15 +5,15 @@ import bio.terra.datarepo.client.ApiClient;
 import bio.terra.datarepo.client.ApiException;
 import bio.terra.datarepo.model.*;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import utils.BulkLoadUtils;
+import testscripts.baseclasses.SimpleDataset;
+import testscripts.testutils.BulkLoadUtils;
 import utils.DataRepoUtils;
 import utils.KubernetesClientUtils;
 
-public class ScalePodsToZero extends runner.TestScript {
+public class ScalePodsToZero extends SimpleDataset {
   private static final Logger logger = LoggerFactory.getLogger(ScalePodsToZero.class);
 
   /** Public constructor so that this class can be instantiated via reflection. */
@@ -23,7 +23,6 @@ public class ScalePodsToZero extends runner.TestScript {
   }
 
   private int filesToLoad;
-  private BulkLoadUtils bulkLoadUtils;
 
   public void setParameters(List<String> parameters) throws Exception {
     if (parameters == null || parameters.size() == 0) {
@@ -34,20 +33,17 @@ public class ScalePodsToZero extends runner.TestScript {
     }
   }
 
-  public void setup(Map<String, ApiClient> apiClients) throws Exception {
-    bulkLoadUtils = new BulkLoadUtils();
-    bulkLoadUtils.bulkLoadSetup(apiClients, billingAccount);
-  }
-
   // The purpose of this test is to have a long-running workload that completes successfully
   // while we scale pods to zero and then scale them back up.
   public void userJourney(ApiClient apiClient) throws Exception {
     RepositoryApi repositoryApi = new RepositoryApi(apiClient);
 
     // set up and start bulk load job
-    BulkLoadArrayRequestModel arrayLoad = bulkLoadUtils.buildBulkLoadFileRequest(filesToLoad);
+    BulkLoadArrayRequestModel arrayLoad =
+        BulkLoadUtils.buildBulkLoadFileRequest(
+            filesToLoad, billingProfileModel.getId(), datasetSummaryModel.getId());
     JobModel bulkLoadArrayJobResponse =
-        repositoryApi.bulkFileLoadArray(bulkLoadUtils.getDatasetId(), arrayLoad);
+        repositoryApi.bulkFileLoadArray(datasetSummaryModel.getId(), arrayLoad);
 
     // =========================================================================
     /* Manipulating kubernetes pods during file ingest */
@@ -96,11 +92,6 @@ public class ScalePodsToZero extends runner.TestScript {
     // =========================================================================
 
     // wait for the job to complete and print out results
-    bulkLoadUtils.getAndDisplayResults(repositoryApi, bulkLoadArrayJobResponse);
-  }
-
-  public void cleanup(Map<String, ApiClient> apiClients) throws Exception {
-    KubernetesClientUtils.changeReplicaSetSizeAndWait(1);
-    bulkLoadUtils.cleanup(apiClients);
+    BulkLoadUtils.getAndDisplayResults(repositoryApi, bulkLoadArrayJobResponse);
   }
 }
